refactor(pocket3d): share skin bindings and isolate mesh ownership
