+++ .github/workflows/pocket3d-handheld.yml
@@ -1,10 +1,10 @@
 name: Pocket3D handheld
 on:
   pull_request:
-    paths: ['engine/pocket3d/crates/pocket3d-anim/**', 'engine/pocket3d/crates/pocket3d/**', 'engine/pocket3d/backends/citro3d/**', 'engine/Cargo.toml', 'engine/Cargo.lock', '.github/workflows/pocket3d-handheld.yml']
+    paths: ['engine/pocket3d/crates/pocket3d-anim/**', 'engine/pocket3d/crates/pocket3d-mesh/**', 'engine/pocket3d/crates/pocket3d/**', 'engine/pocket3d/backends/citro3d/**', 'engine/Cargo.toml', 'engine/Cargo.lock', '.github/workflows/pocket3d-handheld.yml']
   push:
     branches: [main]
-    paths: ['engine/pocket3d/crates/pocket3d-anim/**', 'engine/pocket3d/crates/pocket3d/**', 'engine/pocket3d/backends/citro3d/**', 'engine/Cargo.toml', 'engine/Cargo.lock', '.github/workflows/pocket3d-handheld.yml']
+    paths: ['engine/pocket3d/crates/pocket3d-anim/**', 'engine/pocket3d/crates/pocket3d-mesh/**', 'engine/pocket3d/crates/pocket3d/**', 'engine/pocket3d/backends/citro3d/**', 'engine/Cargo.toml', 'engine/Cargo.lock', '.github/workflows/pocket3d-handheld.yml']
 permissions:
   contents: read
 concurrency:
@@ -18,14 +18,16 @@ jobs:
       - uses: actions/checkout@v7
       - uses: dtolnay/rust-toolchain@1.97.0
         with:
-          components: clippy
+          components: rustfmt, clippy
       - uses: Swatinem/rust-cache@v2
         with:
           workspaces: engine
-      - name: Shared animation and desktop API compatibility
-        run: cargo test --locked --manifest-path engine/Cargo.toml -p pocket3d-anim -p pocket3d --lib
-      - name: Lint shared animation
-        run: cargo clippy --locked --manifest-path engine/Cargo.toml -p pocket3d-anim --all-targets -- -D warnings
+      - name: Shared animation and mesh contracts
+        run: cargo test --locked --manifest-path engine/Cargo.toml -p pocket3d-anim -p pocket3d-mesh --lib
+      - name: Format portable crates
+        run: cargo fmt --manifest-path engine/Cargo.toml -p pocket3d-anim -p pocket3d-mesh -- --check
+      - name: Lint portable crates
+        run: cargo clippy --locked --manifest-path engine/Cargo.toml -p pocket3d-anim -p pocket3d-mesh --all-targets -- -D warnings
       - name: Build independent PICA fixture
         run: >-
           docker run --rm -v "$PWD:/repo"
