+++ engine/Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "crates/pocket-widget",
     "pocket3d/crates/pocket3d",
     "pocket3d/crates/pocket3d-anim",
+    "pocket3d/crates/pocket3d-mesh",
     "pocket3d/crates/pocket3d-bsp",
     "pocket3d/crates/pocket3d-cook",
     "pocket3d/crates/pocket3d-world",
@@ -49,6 +50,7 @@ repository = "https://github.com/pocket-stack/pocketjs"
 [workspace.dependencies]
 pocket3d = { path = "pocket3d/crates/pocket3d" }
 pocket3d-anim = { path = "pocket3d/crates/pocket3d-anim" }
+pocket3d-mesh = { path = "pocket3d/crates/pocket3d-mesh" }
 pocket3d-bsp = { path = "pocket3d/crates/pocket3d-bsp" }
 pocket3d-world = { path = "pocket3d/crates/pocket3d-world" }
 pocket-db = { path = "crates/pocket-db" }
