+++ engine/pocket3d/crates/pocket3d-mesh/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "pocket3d-anim"
-description = "Pocket3D skeletal animation sampling and pose interpolation"
+name = "pocket3d-mesh"
+description = "Portable Pocket3D skin bindings and colored mesh assets"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 
 [dependencies]
+pocket3d-anim = { workspace = true }
 glam = { workspace = true, features = ["libm"] }
